routers refact
