+++ helpers/resources.js
@@ -1,7 +1,9 @@
 "use strict";
 
-const express = require('express');
-const router  = express.Router();
+const express     = require('express');
+const url         = require('url');
+const querystring = require('querystring');
+const router      = express.Router();
 
 module.exports = (knex) => {
   const helpers = require('../helpers/index')(knex);
